BC-1777 WIP: fix module structure
