+++ apps/server/src/console/console.module.ts
@@ -1,8 +1,12 @@
 import { Configuration } from '@hpi-schul-cloud/commons/lib';
-import { Module } from '@nestjs/common';
+import { Dictionary, IPrimaryKey } from '@mikro-orm/core';
+import { MikroOrmModule } from '@mikro-orm/nestjs';
+import { Module, NotFoundException } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
+import { ALL_ENTITIES } from '@shared/domain';
 import { ConsoleWriterModule } from '@shared/infra/console/console-writer/console-writer.module';
 import { KeycloakModule } from '@shared/infra/identity-management/keycloak/keycloak.module';
+import { DB_PASSWORD, DB_URL, DB_USERNAME } from '@src/config';
 import { FilesModule } from '@src/modules/files';
 import { ManagementModule } from '@src/modules/management/management.module';
 import serverConfig from '@src/server.config';
@@ -21,6 +25,19 @@ import { ServerConsole } from './server.console';
 			load: [serverConfig],
 		}),
 		...((Configuration.get('FEATURE_IDENTITY_MANAGEMENT_ENABLED') as boolean) ? [KeycloakModule] : []),
+		MikroOrmModule.forRoot({
+			// TODO repeats server module definitions
+			type: 'mongo',
+			clientUrl: DB_URL,
+			password: DB_PASSWORD,
+			user: DB_USERNAME,
+			entities: ALL_ENTITIES,
+			allowGlobalContext: true,
+			findOneOrFailHandler: (entityName: string, where: Dictionary | IPrimaryKey) => {
+				// eslint-disable-next-line @typescript-eslint/restrict-template-expressions
+				return new NotFoundException(`The requested ${entityName}: ${where} has not been found.`);
+			},
+		}),
 	],
 	providers: [
 		/** add console services as providers */
