+++ apps/server/src/modules/management/management-server.module.ts
@@ -1,0 +1,43 @@
+import { Dictionary, IPrimaryKey } from '@mikro-orm/core';
+import { MikroOrmModule, MikroOrmModuleSyncOptions } from '@mikro-orm/nestjs';
+import { DynamicModule, Module, NotFoundException } from '@nestjs/common';
+import { ALL_ENTITIES } from '@shared/domain';
+import { MongoMemoryDatabaseModule } from '@shared/infra/database';
+import { MongoDatabaseModuleOptions } from '@shared/infra/database/mongo-memory-database/types';
+import { DB_PASSWORD, DB_URL, DB_USERNAME } from '@src/config';
+import { ManagementModule } from './management.module';
+
+export const defaultMikroOrmOptions: MikroOrmModuleSyncOptions = {
+	findOneOrFailHandler: (entityName: string, where: Dictionary | IPrimaryKey) => {
+		// eslint-disable-next-line @typescript-eslint/restrict-template-expressions
+		return new NotFoundException(`The requested ${entityName}: ${where} has not been found.`);
+	},
+};
+
+@Module({
+	imports: [
+		ManagementModule,
+		MikroOrmModule.forRoot({
+			...defaultMikroOrmOptions,
+			// TODO repeats server module definitions
+			type: 'mongo',
+			clientUrl: DB_URL,
+			password: DB_PASSWORD,
+			user: DB_USERNAME,
+			entities: ALL_ENTITIES,
+		}),
+	],
+})
+export class ManagementServerModule {}
+
+@Module({
+	imports: [ManagementModule, MongoMemoryDatabaseModule.forRoot({ ...defaultMikroOrmOptions })],
+})
+export class ManagementServerTestModule {
+	static forRoot(options?: MongoDatabaseModuleOptions): DynamicModule {
+		return {
+			module: ManagementModule,
+			imports: [MongoMemoryDatabaseModule.forRoot({ ...defaultMikroOrmOptions, ...options })],
+		};
+	}
+}
