+++ apps/server/src/modules/management/management.module.spec.ts
@@ -1,7 +1,8 @@
 import { Test, TestingModule } from '@nestjs/testing';
 
 import { DatabaseManagementController } from './controller/database-management.controller';
-import { ManagementModule, ManagementTestModule } from './management.module';
+import { ManagementServerTestModule } from './management-server.module';
+import { ManagementModule } from './management.module';
 import { DatabaseManagementUc } from './uc/database-management.uc';
 
 describe('ManagementModule', () => {
@@ -35,7 +36,7 @@ describe('ManagementTestModule', () => {
 
 	beforeAll(async () => {
 		module = await Test.createTestingModule({
-			imports: [ManagementTestModule],
+			imports: [ManagementServerTestModule],
 		}).compile();
 
 		service = module.get(DatabaseManagementUc);
