+++ apps/server/test/modules/management/database.e2e-spec.ts
@@ -2,7 +2,7 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { INestApplication } from '@nestjs/common';
 import request from 'supertest';
 import { MikroORM } from '@mikro-orm/core';
-import { ManagementTestModule } from '@src/modules/management/management.module';
+import { ManagementServerTestModule } from '@src/modules/management/management-server.module';
 
 describe('Database Management Controller (e2e)', () => {
 	let app: INestApplication;
@@ -11,7 +11,7 @@ describe('Database Management Controller (e2e)', () => {
 
 	beforeAll(async () => {
 		const module: TestingModule = await Test.createTestingModule({
-			imports: [ManagementTestModule],
+			imports: [ManagementServerTestModule],
 		}).compile();
 
 		app = module.createNestApplication();
